+++ src/Elsa.CustomActivities.Tests/Activities/Shared/CustomQuestionTests.cs
@@ -3,6 +3,7 @@ using Elsa.ActivityResults;
 using Elsa.CustomActivities.Activities.Shared;
 using Elsa.CustomModels;
 using Elsa.Services.Models;
+using He.PipelineAssessment.Common.Tests;
 using Xunit;
 
 namespace Elsa.CustomActivities.Tests.Activities.Shared
