+++ src/Elsa.CustomWorkflow.Sdk.Tests/HttpClients/ElsaServerHttpClientTests.cs
@@ -1,6 +1,7 @@
 ﻿using AutoFixture.Xunit2;
 using Elsa.CustomWorkflow.Sdk.HttpClients;
 using Elsa.CustomWorkflow.Sdk.Models.Workflow;
+using He.PipelineAssessment.Common.Tests;
 using Moq;
 using System.Net;
 using Xunit;
