+++ src/Elsa.Server.Tests/Elsa.Server.Tests.csproj
@@ -32,6 +32,7 @@
 
 	<ItemGroup>
 		<ProjectReference Include="..\Elsa.Server\Elsa.Server.csproj" />
+		<ProjectReference Include="..\He.PipelineAssessment.Common.Tests\He.PipelineAssessment.Common.Tests.csproj" />
 	</ItemGroup>
 
 </Project>
