+++ src/Elsa.Server.Tests/Features/Workflow/SaveAndContinue/SaveAndContinueCommandHandlerTests.cs
@@ -9,6 +9,7 @@ using Elsa.Server.Features.Workflow.SaveAndContinue;
 using Elsa.Server.Models;
 using Elsa.Services;
 using Elsa.Services.Models;
+using He.PipelineAssessment.Common.Tests;
 using Moq;
 using Xunit;
