+++ src/Elsa.Server.Tests/Features/Workflow/SaveAndContinue/SaveAndContinueMapperTests.cs
@@ -2,6 +2,7 @@
 using Elsa.CustomModels;
 using Elsa.Server.Features.Workflow.SaveAndContinue;
 using Elsa.Server.Providers;
+using He.PipelineAssessment.Common.Tests;
 using Moq;
 using Xunit;
