+++ src/Elsa.Server.Tests/Features/Workflow/StartWorkflow/StartWorkflowCommandHandlerTests.cs
@@ -6,6 +6,7 @@ using Elsa.Server.Features.Workflow.StartWorkflow;
 using Elsa.Server.Models;
 using Elsa.Services;
 using Elsa.Services.Models;
+using He.PipelineAssessment.Common.Tests;
 using Moq;
 using Xunit;
