+++ src/He.PipelineAssessment.Common.Tests/AutoMoqDataAttribute.cs
@@ -1,15 +1,18 @@
 using AutoFixture;
 using AutoFixture.AutoMoq;
+using AutoFixture.Kernel;
 using AutoFixture.Xunit2;
+using Microsoft.AspNetCore.Mvc;
 
-namespace Elsa.Server.Tests
+namespace He.PipelineAssessment.Common.Tests
 {
     public class AutoMoqDataAttribute : AutoDataAttribute
     {
         public AutoMoqDataAttribute() :
             base(() => new Fixture()
                 .Customize(new AutoMoqCustomization())
-                .Customize(new OmitOnRecursionFixtureCustomization()))
+                .Customize(new OmitOnRecursionFixtureCustomization())
+                .Customize(new AspNetCustomization()))
         {
         }
     }
@@ -21,6 +24,28 @@ namespace Elsa.Server.Tests
             fixture.Behaviors.OfType<ThrowingRecursionBehavior>()
                 .ToList().ForEach(b => fixture.Behaviors.Remove(b));
             fixture.Behaviors.Add(new OmitOnRecursionBehavior(1));
+        }
+    }
+
+    public class AspNetCustomization : ICustomization
+    {
+        public void Customize(IFixture fixture)
+        {
+            fixture.Customizations.Add(new ControllerBasePropertyOmitter());
+        }
+    }
+
+    public class ControllerBasePropertyOmitter : Omitter
+    {
+        public ControllerBasePropertyOmitter()
+            : base(new OrRequestSpecification(GetPropertySpecifications()))
+        {
+        }
+
+        private static IEnumerable<IRequestSpecification> GetPropertySpecifications()
+        {
+            return typeof(Controller).GetProperties().Where(x => x.CanWrite)
+                .Select(x => new PropertySpecification(x.PropertyType, x.Name));
         }
     }
