+++ src/He.PipelineAssessment.Data/SinglePipeline/EsriSinglePipelineClient.cs
@@ -2,11 +2,6 @@
 
 namespace He.PipelineAssessment.Data.SinglePipeline
 {
-    public interface IEsriSinglePipelineClient
-    {
-        Task<string?> GetSinglePipelineData(string spid);
-    }
-
     public class EsriSinglePipelineClient : IEsriSinglePipelineClient
     {
