+++ src/He.PipelineAssessment.Data/SinglePipeline/IEsriSinglePipelineClient.cs
@@ -1,0 +1,7 @@
+﻿namespace He.PipelineAssessment.Data.SinglePipeline
+{
+    public interface IEsriSinglePipelineClient
+    {
+        Task<string?> GetSinglePipelineData(string spid);
+    }
+}
