+++ src/He.PipelineAssessment.UI.Tests/Features/Workflow/SaveAndContinue/SaveAndContinueCommandHandlerTests.cs
@@ -2,6 +2,7 @@
 using Elsa.CustomWorkflow.Sdk.HttpClients;
 using Elsa.CustomWorkflow.Sdk.Models;
 using Elsa.CustomWorkflow.Sdk.Models.Workflow;
+using He.PipelineAssessment.Common.Tests;
 using He.PipelineAssessment.UI.Features.Workflow.SaveAndContinue;
 using Moq;
 using Xunit;
