+++ src/He.PipelineAssessment.UI.Tests/He.PipelineAssessment.UI.Tests.csproj
@@ -30,6 +30,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\He.PipelineAssessment.Common.Tests\He.PipelineAssessment.Common.Tests.csproj" />
     <ProjectReference Include="..\He.PipelineAssessment.UI\He.PipelineAssessment.UI.csproj" />
   </ItemGroup>
