+++ src/He.PipelineAssessment.sln
@@ -44,7 +44,13 @@ Project("{E53339B2-1760-4266-BCC7-CA923CBCF16C}") = "docker-compose", "docker-co
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "He.PipelineAssessment.Infrastructure", "He.PipelineAssessment.Infrastructure\He.PipelineAssessment.Infrastructure.csproj", "{7F47C149-E1A3-4E02-8C24-A4BAF3FF63D0}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "He.PipelineAssessment.Data", "He.PipelineAssessment.Data\He.PipelineAssessment.Data.csproj", "{00052545-39E5-482B-AD76-8A663A753EBC}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "He.PipelineAssessment.Data", "He.PipelineAssessment.Data\He.PipelineAssessment.Data.csproj", "{00052545-39E5-482B-AD76-8A663A753EBC}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "He.PipelineAssessment.Data.Tests", "He.PipelineAssessment.Data.Tests\He.PipelineAssessment.Data.Tests.csproj", "{C4C045BE-E653-4005-A406-BE3A92989D84}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests.Common", "Tests.Common", "{A3BBE60D-1B99-4D99-B22D-CBD7074FFE69}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "He.PipelineAssessment.Common.Tests", "He.PipelineAssessment.Common.Tests\He.PipelineAssessment.Common.Tests.csproj", "{29D4A6A4-216E-44A1-9AD6-DEB2CD27567E}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -108,6 +114,14 @@ Global
 		{00052545-39E5-482B-AD76-8A663A753EBC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{00052545-39E5-482B-AD76-8A663A753EBC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{00052545-39E5-482B-AD76-8A663A753EBC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C4C045BE-E653-4005-A406-BE3A92989D84}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C4C045BE-E653-4005-A406-BE3A92989D84}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C4C045BE-E653-4005-A406-BE3A92989D84}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C4C045BE-E653-4005-A406-BE3A92989D84}.Release|Any CPU.Build.0 = Release|Any CPU
+		{29D4A6A4-216E-44A1-9AD6-DEB2CD27567E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{29D4A6A4-216E-44A1-9AD6-DEB2CD27567E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{29D4A6A4-216E-44A1-9AD6-DEB2CD27567E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{29D4A6A4-216E-44A1-9AD6-DEB2CD27567E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -118,6 +132,8 @@ Global
 		{60B5792C-6C41-4BED-AF95-27078ED13405} = {215A0830-F4AB-44AB-BED3-AA170F692B51}
 		{FEFA3943-09A7-4201-A34D-4CC4B36243D9} = {D9C1B281-DBBB-46D1-AC59-2AF021AC8FAE}
 		{A1D45CCE-7090-4F14-ACD0-A3F5C7DFCBB7} = {04B38B6E-B763-47BC-B3F2-7B963C183720}
+		{C4C045BE-E653-4005-A406-BE3A92989D84} = {D9C1B281-DBBB-46D1-AC59-2AF021AC8FAE}
+		{29D4A6A4-216E-44A1-9AD6-DEB2CD27567E} = {A3BBE60D-1B99-4D99-B22D-CBD7074FFE69}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {8FBFE024-7094-4738-B42D-4214E6140235}
